05-capstone/requirements.txt
alembic==1.4.0
aniso8601==8.0.0
certifi==2019.9.11
Click==7.0
ecdsa==0.15
Flask==1.1.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.1
MarkupSafe==1.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycryptodome==3.9.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
rsa==4.7
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==1.0.0
WTForms==2.2.1
